requirements.txt
confluent-kafka==2.3.0
faker==24.9.0
pydantic==2.7.4
pytest==8.2.0
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9
boto3==1.34.0
